+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 - Post-quantum cryptography is now enabled by default
+- [Feature] Add functions to get default VPN settings and free allocated resources.
+    - See `vpn_get_default_settings` and `vpn_free_default_settings`.
 
 ## 0.99.63